BTL0.3

+++ src/Javacore_BTL/IProduct.java
@@ -1,7 +1,7 @@
 package Javacore_BTL;
 
 public interface IProduct {
-	public static final float MIN_INTEREST_RATE = (float) 0.2;
+	public static final Double MIN_INTEREST_RATE = (Double) 0.2;
 	public void inputData();
 	public void displayData();
 	public void calProfit();
